+++ django_countries/tests/settings_noi18n.py
@@ -1,0 +1,3 @@
+from .settings import *
+
+USE_I18N = False